testing standard build for `NetworkBuilder`
